move tests out of __tests__ folders
