+++ src/book/Page.test.js
@@ -1,4 +1,4 @@
-import Page from '../Page';
+import Page from './Page';
 
 let page;
 beforeEach(() => {
