--- src/dom-utils/__tests__/prefixClass.test.js
@@ -1,9 +1,0 @@
-import { c } from '../index';
-
-test('Adds prefix', () => {
-  expect(c('classname')).toEqual('📖-classname');
-});
-
-test('Adds prefix with dot', () => {
-  expect(c('.classname')).toEqual('.📖-classname');
-});
