+++ src/dom-utils/prefixer.test.js
@@ -1,0 +1,9 @@
+import prefix from './prefixer';
+
+test('Adds prefix', () => {
+  expect(prefix('classname')).toEqual('📖-classname');
+});
+
+test('Adds prefix with dot', () => {
+  expect(prefix('.classname')).toEqual('.📖-classname');
+});
